Try a new SimpleHttpClient
This has a smaller NullClient than before, and better maps to James's article.
The entire old version is still in infrastructure/mod.rs, but this thinner
wrapper should make it easier to also create an HttpClient on top that logs the
requests for testing.

I'm trying to use the Thin Wrapper and Output Tracking patterns, although I only
have Thin Wrapper here.