Test http::Client request
This does one targeted integration test on the Client in http. It's a lot of
work because spawning a webserver in a test and calling it was a challenge, but
I finally just used `http/tests/server.rs` from the reqwest repository to create
a server.

The integrated test server just echos back what is sent to it, and brought in
serde_json and hyper to the app (as dev dependencies). I also brought in anyhow,
because dealing with errors was getting annoying again.

The errors aren't yet mapped on the client, and I have a ton of cleanup to do
but I can finally start working backwards from client to the app.