Get the request/response working with a bearer token and an HTML response
In order to properly close the pop-up window and pass the token we need to return HTML from this response and call post-message, so this function
now returns a small amount of JavaScript. In addition the oauth/token response is not just the token but a JSON response which this correctly parses.